requirements.txt
asgiref==2.3.0
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.5.1
Automat==0.6.0
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
channels==2.1.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
daphne==2.1.1
Django==2.0.5
django-admin-honeypot==1.1.0
django-appconf==1.0.2
django-compressor==2.2
google-api-core==1.1.2
google-auth==1.4.1
google-auth-oauthlib==0.2.0
google-cloud-bigquery==0.31.0
google-cloud-core==0.28.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
oauthlib==2.0.7
priority==1.3.0
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==18.0.0
python-memcached==1.59
pytz==2018.3
rcssmin==1.0.6
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
rsa==3.4.2
service-identity==17.0.0
six==1.11.0
Twisted==18.4.0
txaio==2.10.0
urllib3==1.22
whitenoise==3.3.1
zope.interface==4.5.0
